add systemd service to reboot if a different IP addr is provided by the DHCP server
Sometimes, the DHCP server does not provide the local static IP
requested. That fucks up the SSH aliases that I have. This happens even
though said IP is not used by another machine. This is a temporary
measure to ensure that a given machine has the expected IP address when
trying to remotely access it via SSH, until I get a "real" router.