move the latest stable kernel back to the non-hardened variant
message from nix:
error: The attribute `linuxPackages_hardened_latest' was dropped because the hardened patches
frequently lag behind the upstream kernel. In some cases this meant that this attribute
had to refer to an older kernel[1] because the latest hardened kernel was EOL and
the latest supported kernel didn't have patches.

If you want to use a hardened kernel, please check which kernel minors are supported
and use a versioned attribute, e.g. `linuxPackages_5_10_hardened'.

[1] for more context: https://github.com/NixOS/nixpkgs/pull/133587