scripts: nix-ci: builder: s/native-only/use-emulation/g
Previously, if a given system had support for emulating non-native
systems using QEMU's user emulation provided via Linux kernels' binfmt
support, that system type would be considered supported. Meaning, if an
`x86_64-linux` system could emulate a binary targeted for
`aarch64-linux`, both of these systems' Nix build targets would be
built.

This was nice when I was testing locally. But this default behaviour is
undesired. Therefore, change the behaviour to not search for system's
emulated capabilities and rather let the user explicitly specify it
using the flag that replaces this: `--use-emulation`.