nix.conf: enable binary cache fallback
There can be times when one more Nix binary cache is unreachable. In
that case, Nix just fails. This is because Nix will only query another
binary cache if the first binary cache queried returns a 404, not when
it cannot even be reached. This issue in upstream is in an absolute
dogshit state and I need to use this workaround, which might cause a
rebuild, but is a solution than everything failing.