bump LTS kernel version to v6.12
The v6.12 kernel is the last stable kernel release of 2024 and as a
result will be maintained as an LTS kernel [0]. OpenZFS has released
ZFS v2.2.7 to support the v6.12 kernel [1]. Both master [2] and
stable [2] (which is 24.11 at the time of writing this) branches of
nixpkgs have this version bump for the ZFS kernel module.

It is safe to upgrade the LTS kernel to v6.12.

[0] https://lore.kernel.org/lkml/2024120520-mashing-facing-6776@gregkh/
[1] https://github.com/openzfs/zfs/releases/tag/zfs-2.2.7
[2] https://github.com/NixOS/nixpkgs/commit/e0f2595ed33e60380f8d43695972aad8c9b47be3
[3] https://github.com/NixOS/nixpkgs/commit/79c6b3b11c1765eb873bb2a48979e479eca5dfe4