re-enable systemd in initrd with zfs on root
I mistakenly disabled systemd in initrd when ZFS was used on root where
the initrd could not find the zpool to mount. I believed that this was
because systemd support in initrd was not complete but this was actually
caused by supplying the wrong DTB path which I fixed with 63f7680bbcdd
("systems/chaturvyas: fix wrong extraConfig.dtbRelativePath").