services/user: reset-systemd-user-units: init
This systemd service runs just before shutdown, triggered by the
`ExecStop`, and its only job is to remove all the old systemd units that
home-manager that might have left in the `systemUser`'s
`$HOME/.config/systemd/user` directory. This isn't an issue on NixOS
because `/etc/systemd/system` itself is a link to
`/etc/static/systemd/system` which is a link from the Nix Store.
Meaning, no old services are present in the new system. But
`$HOME/.config/systemd/user` is not a link to anything and only the
files **inside** it are, and the systemd user daemon deals with it
independently of Nix and home-manager.

Also, since this systemd service only runs on NixOS, all units in
`$HOME/.config/systemd/user` are managed by home-manager, **no exceptions**.

Therefore, remove the directory before poweroff/reboot and force
home-manager to re-create with fresh ones, no residue.