services/user: custom-home-manager-upgrade: handle activation exit
The activation script can sometimes fail to exit cleanly and that can
cause systemd to mark the service as failed. Which is _true_ but I don't
want _that_ particular failure to be marked as a failure. I care about
the home-manager configuration not building, not about if it activates
or not, because this isn't NixOS.