flake.nix: add "package": `binfmtCheck`
The `binfmtCheck` package is added to easily check if binary emulation
of another Nix system is working or not, instead of going through the
mess of iterating over files in /proc/sys/fs/binfmt_misc, then realising
that NixOS names the registered binfmts differently than Debian/Ubuntu.

Instead, just try to run it; it works if it runs (with the correct
output string), otherwise binary emulation isn't working.