scripts/builder: only use system(s) provided by `--exclusive-nix-system` flag
I use this for GH actions where each job builds a Nix target for a
single architecture. Noticed that jobs for riscv64-linux would build
for x86_64-linux because the `ci_variables['all_supported_systems']`
dict would be empty and that would trigger the use of determining and
using the native Nix system. This was a logical error. Fixed.