scripts/builder: refactor the `outPath` eval code
Instead of `--link-only` being an independent flag, now it is under
`--evaluate-outPaths` and is renamed to `--link-outPaths`.

Also added a new option called `--github-ci-shortcut` which, instead
of building a Nix expression, just checks the Nix store cache
(the S3 bucket) if the `outPath` for said expression is present or not.

This has the following benefits:
1. Extremely minimal network I/O.
2. Literally zero build time, only Nix evaluation.
3. Extremely minimal network I/O means less hit to my wallet with the
   egress cost from AWS. Ingress if free (in this case). :)
4. Literally zero disk space consumption (other than what is needed for
   the Nix expression evaluation).
5. Reduces the jibber jabber between Nix and the S3 bucket of fetching
   the paths from the remote cache to the disk.