networking: fix "martian source" errors
Sometimes, the `primaryNetIface` likes getting it's own address when the
NixOS system is a VM and also has virtualisation (virtual bridge)
enabled. That happens randomly but is very likely to happen in the first
60 seconds after a getty prompt is reached.

This change makes sure that the IPv4 and the IPv6 addresses are empty
from the NixOS configuration side.