systems/matsya: temporarily switch to mainline kernel
The following not-yet-merged patch necessary for Intel N100 "SoC" is now
merged in the 6.15 kernel which is the current mainline kernel.

> https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/drivers/edac/igen6_edac.c?h=v6.15-rc1&id=267e5b1d267539d9a927dc04aab6f15aca57da92
> EDAC/igen6: Fix the flood of invalid error reports

I was carrying that patch for the `matsya` system. And since it was
merged in Torvald's tree, it is now possible to use the kernel from
nixpkgs on the N100 system. We now avoid rebuilding the kernel on x86.