overlays: comment out qemu overlays for use in future
The QEMU overlays were "removed" because they take a lot of time to
compile, and as I have realized, somewhat unnecessarily. Its primary
purpose was to prevent breakage from bullshit from Python tooling in
general but it ended up using more CPU cycles than it saved. Time to
retire it.