scripts/builder: run eval in single thread if necessary
On hosts with memory less than 8GB, the multi-threaded
`nix eval ${expr}.outPath` evaluations result in a machine hang _with
working ping_. When memory is less than 8GB, the evaluation is
single-threaded.

It is possible to **enforce** single threaded evaluation using the
`--single-thread-eval` option. But it is not possible to do the
opposite, force multi-threaded evaluation, for obvious reasons. Not
that it is not possible but rather I have not allowed it to be possible.