reduce the amount of nixpkgs (plus home-manager) inputs
4 nixpkgs channels and 2 home-manager channels were being tracked:
- nixos-stable
- nixos-stable-small
- nixos-unstable
- nixos-unstable-small
- home-manager-stable
- home-manager-unstable

There were 2 additional home-manager inputs for tracking the -small
channels of nixpkgs. Bringing the number of inputs to 8. This was rather
unnecessary because realistically, only nixos-unstable and
home-manager-unstable were being used.

On top of that, the preferred method of consuming `pkgs` from nixpkgs is
to use the `nixpkgs.legacyPackages.${system}` attribute set, instead of
`import`-ing nixpkgs.