introduce attr `nixosSystemConfig.coreConfig.addrMAC`
This commit introduces the `nixosSystemConfig.coreConfig.addrMAC`
attribute for the primary purpose of binding a MAC address to a specific
IP address in my home LAN for static IP.

Some systems have it set to `null`. One of them is my laptop, which is
intentional and other 3 aren't operational daily and will be populated
once that is the case.